+++ phpmyfaq/admin/assets/src/configuration/upgrade.js
@@ -22,7 +22,7 @@ export const handleCheckForUpdates = () => {
   if (button) {
     button.addEventListener('click', (event) => {
       event.preventDefault();
-      fetch('index.php?action=ajax&ajax=updates&ajaxaction=check-updates', {
+      fetch('/admin/api/update-check', {
         method: 'POST',
         headers: {
           Accept: 'application/json, text/plain, */*',