+++ phpmyfaq/admin/record.edit.php
@@ -15,6 +15,7 @@
  * @since     2003-02-23
  */
 
+use phpMyFAQ\Administration\AdminLog;
 use phpMyFAQ\Attachment\AttachmentFactory;
 use phpMyFAQ\Category;
 use phpMyFAQ\Category\CategoryRelation;
@@ -28,7 +29,6 @@ use phpMyFAQ\Helper\CategoryHelper;
 use phpMyFAQ\Helper\LanguageHelper;
 use phpMyFAQ\Helper\UserHelper;
 use phpMyFAQ\Link;
-use phpMyFAQ\AdminLog;
 use phpMyFAQ\Question;
 use phpMyFAQ\Revision;
 use phpMyFAQ\Session\Token;